- use size_t, ssize_t, off_t wherever appropriate - push casts to "int" as far as possible down the code (e.g. down to Tcl_* functions accepting sizes, but defined with "int", such as Tcl_WriteChars, Tcl_NewStringObj, Tcl_NewByteArrayObj) - be more careful on redefining macros (MSC seems a moving target) - provide base definition for inttypes for MSC WIN32 and WIN64 - use NS_SOCKET instead of SOCKET (which is defined under Windows). We should stick with macros in our own namespace whenever possible. - added the project files and the solution file for visual studio

+++ nsd/adpeval.c
@@ -317,6 +317,7 @@ void
 NsAdpInit(NsInterp *itPtr)
 {
     Tcl_DStringInit(&itPtr->adp.output);
+
     NsAdpReset(itPtr);
 }
 
@@ -1028,7 +1029,7 @@ AdpExec(NsInterp *itPtr, int objc, Tcl_Obj *objv[], char *file,
     frame.ident = NULL;
     savecwd = itPtr->adp.cwd;
     if (file != NULL && (slash = strrchr(file, '/')) != NULL) {
-        Ns_DStringNAppend(&cwd, file, slash - file);
+      Ns_DStringNAppend(&cwd, file, (int)(slash - file));
         itPtr->adp.cwd = cwd.string;
     }
     frame.prevPtr = itPtr->adp.framePtr;
